requirements.txt
cffi==1.14.3
config==0.3.9
confusable-homoglyphs==3.2.0
cryptography==3.2.1
dj-database-url==0.5.0
Django==1.11.17
django-bootstrap3==12.0.0
django-countries==6.1.3
django-heroku==0.3.1
django-registration==2.4.1
django-rest-knox==4.1.0
djangorestframework==3.8.2
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
python-decouple==3.3
pytz==2020.4
six==1.15.0
whitenoise==5.2.0
